requirements.txt
python-dotenv>=0.20.0
google-cloud-storage>=1.33.0
earthengine-api>=0.1.260